requirements.txt
Pillow>=2.5.1
requests>=2.3.0
beautifulsoup4>=4.3.2
boto>=2.38.0
pymongo>=2.6.3
